requirements.txt
streamlit
numpy
Pillow
opencv-python-headless
tensorflow
ultralytics
torch
torchvision
